requirements.txt
streamlit>=1.32
groq>=0.11
python-dotenv>=1.0
pandas>=2.1
vaderSentiment>=3.3.2

langdetect>=1.0.9
